Conflicts:
	data/json/items/ammo.json
	data/json/items/mods.json
	data/json/items/ranged.json
	src/ranged.cpp

Conflicts:
	data/json/items/ammo.json
	data/json/items/archery.json
	data/json/items/armor.json
	data/json/items/bionics.json
	data/json/items/books.json
	data/json/items/comestibles.json
	data/json/items/melee.json
	data/json/items/mods.json
	data/json/items/ranged.json
	data/json/items/tools.json
	data/mods/EW_Pack/ew_ammo.json
	data/mods/EW_Pack/ew_books.json
	data/mods/EW_Pack/ew_weapons.json
	src/iuse.cpp
	src/uistate.h

+++ src/iuse.cpp
@@ -4550,7 +4550,7 @@ int iuse::turret(player *p, item *, bool)
 
  p->moves -= 100;
  monster mturret(GetMType("mon_turret"), dirx, diry);
- int ammo = std::min(p->inv.charges_of("9mm"), 500);
+ int ammo = std::min(p->inv.charges_of("9mm"), long(500));
  if (ammo > 0) {
     p->inv.reduce_charges(p->inv.position_by_type("9mm"), ammo);
     if (ammo == 1) {
